[SystemInfo.py] Switch from kernel module to text file
This change swaps from using the new kernel module to a text file representation of the kernel module values.  This change is being made to improve the acceptability of the BoxInfo code for other images.

The new "enigma.info" file is generated by the build system and contains exactly the same data as provided by the kernel module.  The new file is located as "/usr/lib/enigma.info" and will thus be far easier to locate than the kernel module whose location varies depending on the kernel in use.  The contents of the "enigma.info" have a checksum that is verified when Enigma2 starts to check the validity / consistency of the file.  This file has been made text to simplify its use and has *NOT* been done to encourage user editing.  PLEASE DO NOT EDIT THIS FILE!  It is created automatically by the build system and maintained by system updates.

This change also adds some new log messages to confirm the integrity of the "enigma.info" file as well as other status information.