Refactor the MultiBoot tool
* Refactor the MultiBoot tool

- Rename Multiboot.py to MultiBoot.py to be consistent with other usages.
- Replace all references to Multiboot to MultiBoot in all code and strings.

There will be an equivalent change to Vision Core to match these changes.

NOTE: Due to an issue in my Git client I can't change the case of a filename.  MultiBoot.py has to be committed as MultiBootNew.py and I will rename it on the next commit.

* [MultiBoot.py] Correct file name.

See previous commit.