[Information.py] Properly remove the PIL library
This change fully removed the PIL library and the code that uses it.  (The previous change resulted in a code crash.)  The same effect will soon be available via enhancements to the enigma binary and matching changes to skin.py.  A new skin attribute will now allow image scaling with multiple new flags to control aspect ratio and image position.