[config.py] Refactor config.py
* [config.py] Refactor config.py

This change incorporates a significant refactor of config.py with matching changes in ConfigList.py.  The primary reason for this refactor is to stop config.py from triggering a significant number of change notifications and notifier callbacks when they were not appropriate.  The intention is to improve the performance of the entire Setup and config functionality of Enigma2.

Some of the highlights of this change are:
- Stop all config classes from signalling changes and triggering notification when the data does not change.  Navigating around the data is *not* an edit.
- Use callbacks from ConfigList.py to more efficiently allow config.py to notify ConfigList.py that data has changed during an edit.
- Tidy and optimise the code.
- Start camelCasing and improving variable names.
- Add more argument validation code.
- Start removing assert statements and replace them with raise statements so that errors in the code can be more readily detected.
- Sort the classes into groups base on main class followed by the sub classes (in alphabetical order).
- Sort the methods in the classes to a standardised order.
- Improve Python 3 compatibility.

Please note that this is a significant change and is still a work in progress.  Some diagnostic comments are still in the code and will be cleaned up in a later commit.  Please report any bugs or issues so that they can be addressed.  This version is being committed now as initial private testing appears to be going well and we would like to expose the code for wider testing in a larger community.  If significant issues are found that can not be corrected within hours or reports being submitted this commit may be reverted until such time that the critical issues are addressed.

This code written by IanSav with the assistance of Jbleyel.

* [ConfigList.py] Align with new config.py

* [config.py] More improvements

- Correct arguments for getMulti() in ConfigElement.
- Improve ConfigBoolean description initialisation.

* [config.py] Remove a long standing limitation

A ConfigList entry with only one item, a description, is actually valid to the underlying C++ code.  This creates a comment entry in the ConffigList screen.  The entry can not be selected and will be skipped during navigation.

* [ConfigList.py] Don't process comment entries

This change aligns with the restoration of the ConfigList comment feature where a ConfigList entry with only one item, a description, is actually valid to the underlying C++ code. This creates a comment entry in the ConffigList screen. The entry can not be selected and will be skipped during navigation.