[Directories.py] Replace slow stack() call
This commit replaces the inspect.stack() method call with the sys._getframe() method call that is faster yet yields the same results.

On OpenPLi I I have been berated for writing horrible code (and many other things) that contributed to a 35 second delay to OpenPLi starting.  I won't go into details here but the claims a pure rubbish!  (I can provide proof if required.)  However the stack() call did prove to be slower and could be improved so this code update is warranted.

Thanks to @nautilus7 for bringing the issue to my attention.