FlashImage.py: add exception clause to remove tmp file
we had a vulnerability, with the merge of changes for py3 an exception clause was already added for doreReboot, since the STARTUP file must be present on the device
This problem can be generated if the module was used and then you want to restart the system.

now we found another vulnerability:
If you cancel action and remove dir tmp,  command is executed and the file does not exist we have a crash.
the FileNotFoundError clause will avoid this crash.