Remove hard coded enableWrapAround
Now that the skin "windowstyle" allows for the skin's default of enableWrapAround to be set it is no longer necessary to hard code the default.  The system default has been adjusted to set the system default to enabled.  The skin default can be set in the "listbox" tag in the "windowstyle" block of the skin.  As before individual screens can still have an "enableWrapAround " attribute to modify the settings for any individual screen.