SkinSelector.py: changes in list and description for skin display
The list changes, it is not intuitive at all, for example: If we have an XML file for display we would see Default (current),
if we add another type of XML, for example an XML with picon, then we would see skin_default (Default),
It is not intuitive to know what you are going to activate or maintain, preview can help us (but remember that not all display XML files have preview attached).
The description was also wrong, include "None" the display file to activate or keep.

From this change:

We are going to see in the list the XML file in question that we activate, we must take into account that the files are classified in their names,
and on this I base myself to make it more intuitive.
if it is an XML with picon then now we are going to see skin_display_picon (current or default) , if it is not then skin_display (current or default) and so on...
the description reminds us of the file and extension in question that we activate or maintain.

The pending status is updated.

The grautec XML type file is included in the code.