RecordTimer.py: fix integer for timer repetead and fix bug begTime
Fix this bug:
TypeError: in method 'eEPGCache_lookupEventTime', argument 3 of type 'time_t'

C++ time_t in queryTime already starts begTime, the time sequence to start the event time.

Fix this problem:

TimerEntry does not detect booleans for repeating timers, they are integers to sort for repeating weekly, daily, Mon-Fry, or user-defined.

The consequence is that if we restart the UI, it will lose the assigned data, being assigned to an else in TimerEntry ->Weekly.

here it was changed to boolean:

https://github.com/OpenVisionE2/enigma2-openvision/commit/2395ff9498305e5bb699fd436d47c8742c007757