Add Time Wizard Screen
TimeWizard screen:
A new screen is added to configure the local time through Geolocation in the wizard.
the interface until now in the wizard to configure the local time, worked correctly although it has a rather poor interface.
- Added yellow button to set local time.
- The text for the actions with the yellow button is set, it changes state to inform new status.

Wizard:
DEBUG code is cleaned.
Silent DEBUG type:
<Object Components.Fonts.List.List at 0xb0e62c40>
Components.ConfigList.ConfigList object at 0xb0e3ce50>
The Try statement is removed.
The clause is generated because it attempts an integer for the access levels to identify the argument, when there is not an identifying integer in that argument.
Now defined as argument name.

-Added the Start in LocalWizard wizard, this was removed from the wizard without any meaning, a non-English user is forced to press the red button to change language, now we present it to you in the wizard.

- Fixed initial level access settings in wizard,  was broken, we didn't show any list to check for OV, HDMI-CEC updates or set access levels.

- WizardLanguage: correctly adapt button and action in the wizard.