GraphMultiEpg.py: add button blue AutoTimer
-Added the blue button to launch the AutoTimer plugin directly from the EPG charts.
-The long blue button before "date and time input" is now prime time.
- Added TEXT button for "Date and time input".
The skin shows on two separate lines in GraphicalEPG and GraphicalEPGList the actions of the blue button, short press we enter AutoTimer and long press "prime time" or "now".
- Added an explanatory message on how to start a self-timer.
- If the plugin does not exist, we will also display a message.

EPG selection:
The actions of the map are conditioned since if the EPGSearch plugin exists, we enter into conflict with the blue button of the EPG_SINGLE, then:

Channel Selection press button INFO, then:
If EPGSearch is installed as a plugin, the blue button runs EPGSearch to search for EPG SINGLE events (previously we only had the option to select new channel for EPG SINGLE).
If the EPGSearch plugin is not installed, then it will display "Channel Change" to choose a simple EPG from another SimpleChannelSelection service, ie its usual routine action with EPG SINGLE.