PictureInPicture.py: show configurable PiP only if it is service
If there is no service and we try to perform an action on said service, for example REC, the configurable PiP with these measurements and position will appear in the graphic of the error message in the skin, that is, we visualize a small PiP with this measurement - position x , y, w, h = 16, 16, 16, 16, defined in PipPigMode.

So make sure before you create it, that there are PiP service attributes.