InfoBarGenerics.py: not show "Add recording stop after current event" if not exist event
With this we do not show the recording of an event and end after its completion if there is no such event.
Before, recording was indefinite, it does not meet the indication of the item("Add recording (stop after current event") since the recording cannot be canceled since there is no event.

This situation can be generated when we do not have an event, with EIT disabled and without EPG cache and we choose to make an instant recording., so the correct options that should be displayed are: indefinitely, manualendtime and manualduration.