[HelpMenu.py] Expand remote control help diagnostics
Change the HELP screen remote control diagnostics to now include a two line label that provides the key ID name and the expected button label for each button that is displayed during the diagnostic scan performed by the "testHighlights()" method.  (This diagnostic is invoked by long pressing HELP while the help screen is displayed, where the help's help is shown if the button is short pressed.)

Note that this new feature will require a skin change for the new diagnostic to be seen.  As mentioned above, a 'name="label"' widget is required and it should allow for two lines of text.