InfoBarGenerics.py: get AP files in TimeShift
define getFileNameExtensionAP

- Certain services e.g. Eleven Sports 1 4K has properties for .AP files.

With this change we support these files on a timeshift basis saving them if that is the user's intention, this means it can be used for editing with CutListEditor.

- We will display a message when a timeshift with editing properties has been saved to the storage device.

NOTE: Never delete the temporary file timeshift.XXXXXX.AP, if this file belongs to the current event and saved for recording timeshift, it is needed for edits in CutListEditor.