SetupFallbacktuner.py: Added purge channel list
We added the option to purge the list on the server to resolve import issues on the client box.
An incorrect character is the "+" sign in the file name in the channel list (common with the movistar+ provider).
With this change we clean and purge the list to be able to import it.