PowerTimer.py: add info with wake up from EPGImport Plugin
With this we give information with receiver on and we don't know the reason.
If the plugin is not configured to import in standby mode or not go to deep mode after import, the receiver is awake, then we now provide the information plugin performed the activation.

PowerTimer.py: wake up message EPGImport sync with EPGRefresh

- Update width and height MessageBoxSimple

Co-authored-by: norhap <elfogondesegovia@gmail.com>