[ChoiceList.py] Make vertical alignment configurable
Add a new skin parameter called "ChoicelistVerticalAlignment", that takes a value from "*top", "*center", "*middle" or "*bottom", to specify the vertical alignment of all ChoiceList base widgets.  (NOTE: The "*" is required to tell the parameter processor that the data is actually a string value and to not try and process this value through the numeric parsers.)

This parameter is a temporary hack to allow skins from other distributions to work, unchanged, on openATV.  The other images should really fix the alignment issue in their code.  The issue in the other images is that the default alignment is "top" and skinners have been forced to manually calculate offsets of the text to cause the text to be centered vertically in the line.  In openATV we fixed this issue so that no such offset is required.  Manually calculating and maintaining centering offsets is a waste of time and code when Enigma2 is fully capable of centering text itself.  Most Enigma2 displays assume and expect text to be centered in the space allocated.  The opeATV change makes this automatic and consistent for all widgets.